requirements.txt
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lightning-utilities==0.11.9
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-metric-learning==2.8.1
pytz==2024.1
PyYAML==6.0.2
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
supervision==0.22.0
sympy==1.13.2
threadpoolctl==3.5.0
torch==2.4.0
torchaudio==2.4.0
torchmetrics==1.6.0
torchvision==0.19.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
zipp==3.20.0
